Move preprocessFunctionArgumentDefaults out of printer/print.js
